Rename TabSwitcher, update docs and generate snapshots
