--- src/components/TabSwitcher/tests/TabSwitchContainer.js
@@ -1,5 +1,0 @@
-import { generateSnapshots } from 'helpers';
-
-import TabSwitchContainer from '../TabSwitchContainer';
-
-generateSnapshots(TabSwitchContainer);
