+++ src/components/Tabs/tests/TabsWrapper.js
@@ -1,0 +1,5 @@
+import { generateSnapshots } from 'helpers';
+
+import TabsWrapper from '../TabsWrapper';
+
+generateSnapshots(TabsWrapper);
