+++ src/components/index.js
@@ -1,6 +1,6 @@
 // this is the primary export for Components
 // the main export for this lib is `src/index.js`
-export { default as TabSwitcher } from './TabSwitcher';
+export { default as Tabs } from './Tabs';
 export { default as Checkbox } from './Checkbox';
 export { default as Tooltip } from './Tooltip';
 export { default as Note } from './Note';
